Subject: Temporarily deactivate certain build commands and InfluencerAssignmentDialog component
Body: A number of build commands in 'amplify.yml' and the entirety of
the 'InfluencerAssignmentDialog.tsx' component have been commented out
temporarily. These changes are likely related to a troubleshooting
process or a refactor that is currently in progress. It's important to
note that these changes will affect the build process and functionality
of the mentioned component until they're reverted back to their active
states. Further revisions should be expected.

Issue references: N/A